# docs/swagger.yaml
basePath: /admin-api
definitions:
  model.Admin:
    properties:
      created_at:
        type: string
      id:
        type: integer
      nickname:
        type: string
      role:
        $ref: '#/definitions/model.AdminRole'
      role_id:
        type: integer
      status:
        type: integer
      token:
        type: string
      token_expire_time:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.AdminRole:
    properties:
      admins:
        items:
          $ref: '#/definitions/model.Admin'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      see_cb:
        type: integer
      updated_at:
        type: string
    type: object
  request.AdminForm:
    properties:
      id:
        type: integer
      nickname:
        type: string
      password:
        type: string
      role_id:
        type: integer
      status:
        type: integer
      username:
        type: string
    required:
    - nickname
    - password
    - role_id
    - status
    - username
    type: object
  request.AdminRoleForm:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  request.Login:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 后台接口
  title: 供应商后台管理系统API
  version: "1.0"
paths:
  /admin/create:
    post:
      consumes:
      - application/json
      description: 创建管理员
      parameters:
      - description: 管理员信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AdminForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 创建管理员
      tags:
      - 管理员
  /admin/delete:
    post:
      consumes:
      - application/json
      description: 管理员编删除
      parameters:
      - description: 管理员信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AdminForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员删除
      tags:
      - 管理员
  /admin/detail/{id}:
    get:
      consumes:
      - application/json
      description: 管理员详情
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员详情
      tags:
      - 管理员
  /admin/list:
    get:
      consumes:
      - application/json
      description: 管理员列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      - description: 昵称
        in: query
        name: nickname
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员列表
      tags:
      - 管理员
  /admin/update:
    post:
      consumes:
      - application/json
      description: 管理员编辑
      parameters:
      - description: 管理员信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AdminForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员编辑
      tags:
      - 管理员
  /admin_role/create:
    post:
      consumes:
      - application/json
      description: 管理员角色创建
      parameters:
      - description: 管理员角色
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AdminRoleForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员角色创建
      tags:
      - 管理员角色
  /admin_role/delete:
    post:
      consumes:
      - application/json
      description: 管理员角色删除
      parameters:
      - description: 管理员角色信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AdminRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员角色删除
      tags:
      - 管理员角色
  /admin_role/detail/{id}:
    get:
      consumes:
      - application/json
      description: 管理员角色详情
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员角色详情
      tags:
      - 管理员角色
  /admin_role/list:
    get:
      consumes:
      - application/json
      description: 管理员角色列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员角色列表
      tags:
      - 管理员角色
  /admin_role/update:
    post:
      consumes:
      - application/json
      description: 管理员角色编辑
      parameters:
      - description: 管理员信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AdminRoleForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员角色编辑
      tags:
      - 管理员角色
  /file/oss_token:
    get:
      consumes:
      - application/json
      description: 获取ossToken
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 获取ossToken
      tags:
      - 文件
  /file/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传文件到本地
      parameters:
      - description: 上传文件示例
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 上传文件到本地
      tags:
      - 文件
  /index:
    get:
      consumes:
      - application/json
      description: Index
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Index
      tags:
      - Index
  /login:
    post:
      consumes:
      - application/json
      description: 登录
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Admin'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 登录
      tags:
      - 登录
  /logout:
    post:
      consumes:
      - application/json
      description: 登出
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 登出
      tags:
      - 登录
securityDefinitions:
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"